requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-colorfield==0.11.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
flake8==7.0.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
